+++ src/rpg_player.h
@@ -44,7 +44,7 @@ public:
 	bool			key_released;
 	bool			skiped;
 
-	int				nbjoueurs;
+	size_t				nbjoueurs;
 	int				cur_joueur;	// Joueur pris en compte à cet instant (0 ou 1)
 	int				base_joueur[2]; // Représente l'image de base du joueur selon qu'il soit Blip ou Blop
